Move glustercli package under bin
