+++ bin/glustercli/cmd/common.go
