+++ bin/glustercli/cmd/peer.go
